+++ src/main/resources/db/scripts/01_create_cars.sql
@@ -1,5 +1,5 @@
 create table cars(
 	number_plate varchar(8) PRIMARY KEY,
 	model varchar(9) NOT NULL,
-	price_euro int(7)
+	price_euro double(7)
 );
